+++ examples/cttc-nr-tdd-cc-bwp-demo.cc
@@ -481,7 +481,7 @@ main (int argc, char *argv[])
   for (uint32_t j = 0; j < enbNetDev.GetN (); ++j)
     {
       ObjectMapValue objectMapValue;
-      enbNetDev.Get (j)->GetAttribute ("ComponentCarrierMap", objectMapValue);
+      enbNetDev.Get (j)->GetAttribute ("BandwidthPartMap", objectMapValue);
       for (uint32_t i = 0; i < objectMapValue.GetN (); i++)
         {
           Ptr<BandwidthPartGnb> bandwidthPart = DynamicCast<BandwidthPartGnb> (objectMapValue.Get (i));
@@ -651,7 +651,7 @@ main (int argc, char *argv[])
 
   /*
    * To check what was installed in the memory, i.e., BWPs of eNb Device, and its configuration.
-   * Example is: Node 1 -> Device 0 -> ComponentCarrierMap -> {0,1} BWPs -> MmWaveEnbPhy -> MmWavePhyMacCommong-> Numerology, Bandwidth, ...
+   * Example is: Node 1 -> Device 0 -> BandwidthPartMap -> {0,1} BWPs -> MmWaveEnbPhy -> MmWavePhyMacCommong-> Numerology, Bandwidth, ...
   GtkConfigStore config;
   config.ConfigureAttributes ();
   */