+++ examples/cttc-simple-interference-example.cc
@@ -703,7 +703,7 @@ NrSingleBwpSetup::NrSingleBwpSetup (Scenario *scenario, OutputManager *manager,
     {
       ObjectMapValue objectMapValue;
       Ptr<MmWaveEnbNetDevice> netDevice = DynamicCast<MmWaveEnbNetDevice>(m_gnbDev.Get(j));
-      netDevice->GetAttribute("ComponentCarrierMap", objectMapValue);
+      netDevice->GetAttribute("BandwidthPartMap", objectMapValue);
       for (uint32_t i = 0; i < objectMapValue.GetN(); i++)
         {
           Ptr<BandwidthPartGnb> bandwidthPart = DynamicCast<BandwidthPartGnb>(objectMapValue.Get(i));